Attempt to allow for building test files
Standard 'elm make' fails when building test dependencies that import
modules that are in the 'test-dependencies' section of your elm.json
file. In order to make it work you have to use 'elm-test make'.

So we add something that looks for a failure on 'import Test' and
retries the compilation with 'elm-test make' instead. It is a flawed
approach as there might be other test dependencies that the compiler
looks for first and fails on but Test is the most likely by far.

It don't want to rely on a directory check as I currently think it is
possible to write tests outside of the 'tests' directory and just import
them over to that place. I want to leave the option open for writing
tests next to source files. Perhaps we could introduce a flag or
something to switch between this and a directory check if desirable.